image_recog/requirements.txt
efficientnet_pytorch
hypercorn
hypercorn[uvloop]
pyimage
quart
quart-cors
rapidjson
uvloop
